requirements.txt
Pillow>=10.0.0
customtkinter>=5.2.0
tifffile>=2023.1.1
cairosvg>=2.7.0
rawpy>=0.18.0
numpy>=1.24.0
matplotlib>=3.7.0
